fix gallery items being u16 when they could really be negative

+++ src/models/gallery_item.rs
@@ -18,5 +18,5 @@ pub struct GalleryItem {
     created: DateField,
     /// The order of the gallery image.
     /// Gallery images are sorted by this field and then alphabetically by title.
-    ordering: Option<u16>,
+    ordering: Option<i32>,
 }
